Animate on both HTML and BODY because of browsers.
Firefox handles the animation on the HTML node, Chrome handles it
on the BODY node. Safari also handles on the BODY and I'm guessing
IE does on one of them.

Of course, watching both of these nodes creates a scenario where
the callback will fire twice, so we avoid doing anything on HTML.